requirements.txt
django==1.11.4
django-environ==0.4.3
djangorestframework==3.6.3
httpie==0.9.9